refactor: update timist components documentation and restructure Button folder
- Update timist-components-needed.md to reflect actual components needed
- Remove components already covered by existing ones (Alert, Tabs, Drawer, etc.)
- Add Snackbar as completed component replacing Alert and Toast
- Convert all HTML examples to Pug syntax with clean naming conventions
- Remove BEM naming in favor of simple descriptive class names
- Restructure Button folder to Buttons with subfolder structure
- Update all imports to reflect new Button component path
- Total components needed reduced from 13 to 7
